Move all the hooks to hooks folder
