+++ src/components/card-status-dropdown.tsx
@@ -26,8 +26,8 @@ import {
 import { Link } from '@tanstack/react-router'
 import { cn } from '@/lib/utils'
 import { useAuth } from '@/lib/hooks'
-import { useProfile } from '@/lib/use-profile'
-import { useDeckCard } from '@/lib/use-deck'
+import { useProfile } from '@/hooks/use-profile'
+import { useDeckCard } from '@/hooks/use-deck'
 import { buttonVariants } from './ui/button-variants'
 import { Badge } from './ui/badge'
 import { Separator } from './ui/separator'
