+++ src/components/chat/card-preview.tsx
@@ -1,7 +1,7 @@
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
-import { useLanguagePhrase } from '@/lib/use-language'
+import { useLanguagePhrase } from '@/hooks/use-language'
 import Callout from '../ui/callout'
-import { useDeckCard } from '@/lib/use-deck'
+import { useDeckCard } from '@/hooks/use-deck'
 import { uuid } from '@/types/main'
 import { ago } from '@/lib/dayjs'
 import { CardStatusDropdown } from '../card-status-dropdown'
