+++ src/components/friends/pending-requests-header.tsx
@@ -1,5 +1,5 @@
 import { Link } from '@tanstack/react-router'
-import { useRelations } from '@/lib/friends'
+import { useRelations } from '@/hooks/use-friends'
 import { buttonVariants } from '../ui/button-variants'
 import { Badge } from '../ui/badge'
 import { cn } from '@/lib/utils'
