+++ src/components/navs/deck-switcher.tsx
@@ -18,7 +18,7 @@ import {
 	SidebarMenuItem,
 	useSidebar,
 } from '@/components/ui/sidebar'
-import { useProfile } from '@/lib/use-profile'
+import { useProfile } from '@/hooks/use-profile'
 import languages from '@/lib/languages'
 
 const useDeckMenuData = () => {
