+++ src/components/phrase-extra-info.tsx
@@ -1,7 +1,7 @@
 import { CardFull, uuid } from '@/types/main'
 import { ago } from '@/lib/dayjs'
-import { useLanguagePhrase } from '@/lib/use-language'
-import { useDeckCard } from '@/lib/use-deck'
+import { useLanguagePhrase } from '@/hooks/use-language'
+import { useDeckCard } from '@/hooks/use-deck'
 import { dateDiff, intervals, retrievability, roundAndTrim } from '@/lib/utils'
 import Flagged from '@/components/flagged'
 import ExtraInfo from '@/components/extra-info'
