+++ src/components/requests/request-list-item.tsx
@@ -7,7 +7,7 @@ import { avatarUrlify, cn } from '@/lib/utils'
 import { ago } from '@/lib/dayjs'
 import UserPermalink from '@/components/user-permalink'
 import CopyLinkButton from '@/components/copy-link-button'
-import { PhraseRequestFull } from '@/lib/use-requests'
+import type { PhraseRequestFull } from '@/hooks/use-requests'
 import Flagged from '@/components/flagged'
 import { Blockquote } from '@/components/ui/blockquote'
 import { Button } from '@/components/ui/button'
