+++ src/routes/_user/learn.$lang.bulk-add.tsx
@@ -21,7 +21,7 @@ import { Textarea } from '@/components/ui/textarea'
 import ErrorLabel from '@/components/fields/error-label'
 import { ShowAndLogError } from '@/components/errors'
 import languages from '@/lib/languages'
-import { useProfile } from '@/lib/use-profile'
+import { useProfile } from '@/hooks/use-profile'
 import { Separator } from '@/components/ui/separator'
 import { SelectOneOfYourLanguages } from '@/components/fields/select-one-of-your-languages'
 import { LanguageLoaded, PhraseStub } from '@/types/main'
