+++ src/routes/_user/learn.$lang.requests.tsx
@@ -1,4 +1,4 @@
-import { allMyPhraseRequestsQuery } from '@/lib/use-requests'
+import { allMyPhraseRequestsQuery } from '@/hooks/use-requests'
 import { createFileRoute } from '@tanstack/react-router'
 
 export const Route = createFileRoute('/_user/learn/$lang/requests')({
