+++ src/routes/_user/learn.$lang.search.tsx
@@ -13,11 +13,11 @@ import { Input } from '@/components/ui/input'
 import { Button } from '@/components/ui/button'
 import languages from '@/lib/languages'
 import { uuid } from '@/types/main'
-import { useLanguagePhrasesMap, useLanguagePids } from '@/lib/use-language'
+import { useLanguagePhrasesMap, useLanguagePids } from '@/hooks/use-language'
 import { useMemo, type SetStateAction } from 'react'
 import { LanguagePhrasesAccordionComponent } from '@/components/language-phrases-accordion'
 import { FancyMultiSelect } from '@/components/ui/multi-select'
-import { useLanguageTags } from '@/lib/use-language'
+import { useLanguageTags } from '@/hooks/use-language'
 import { Separator } from '@/components/ui/separator'
 import { PhraseSearchSchema, PhraseSearchType } from '@/lib/schemas'
